requirements.txt
click==8.0.3
colorama==0.4.4
document==1.0
Flask==2.0.2
Flask-Cors==3.0.10
Jinja2==3.0.3
python-docx==0.8.11
MarkupSafe==2.0.1
python-docx==0.8.11
six==1.16.0
Werkzeug==2.0.2
docx2pdf==0.1.8
convertapi==1.4.0